requirements.txt
PyYAML==3.11
Pygments==2.1.3
beautifulsoup4==4.4.1
bs4==0.0.1
wheel==0.24.0
wsgiref==0.1.2
